+++ src/main/resources/release_notes.md
@@ -15,6 +15,7 @@ Please do not forget to back up data.
 
 ### Database
 * faster and more secure database access using prepared statements (#1593)
+* refactoring of user administration (#1628)
 
 ### Squad
 * add last match columns: rating at end of match, position and minutes (#1523)